+++ system/platform-core/src/test/java/org/platformlambda/core/MsgPackTest.java
@@ -50,6 +50,9 @@ public class MsgPackTest {
         input.put("boolean", true);
         input.put("array", HELLO_WORLD);
         input.put("integer", 12345L);
+        input.put("long", 12345L);
+        input.put("float", 12.345f);
+        input.put("double", 12.345d);
         input.put("pojo", pojo);
         input.put(PayloadMapper.NOTHING, null);
         byte[] b = msgPack.pack(input);
@@ -57,7 +60,10 @@ public class MsgPackTest {
         assertInstanceOf(Map.class, o);
         Map<String, Object> result = (Map<String, Object>) o;
         // long number will be compressed into integer if applicable
-        assertEquals(Integer.class, result.get("integer").getClass());
+        assertInstanceOf(Integer.class, result.get("integer"));
+        assertInstanceOf(Integer.class, result.get("long"));
+        assertInstanceOf(Float.class, result.get("float"));
+        assertInstanceOf(Double.class, result.get("double"));
         // MsgPack does not transport null elements in a map
         assertFalse(result.containsKey(PayloadMapper.NOTHING));
         result.remove(PayloadMapper.NOTHING);