+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.0.29, 12/3/2024
+
+### Added
+
+Added integer, long, float, double and boolean type matching for state machine.
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
+
 ## Version 4.0.28, 11/29/2024
 
 ### Added